+++ ace-support-core/CHANGELOG.md
@@ -7,6 +7,19 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.14.1] - 2025-12-30
+
+### Changed
+
+* Add ace-config dependency for configuration cascade delegation
+
+## [0.14.0] - 2025-12-30
+
+### Changed
+
+* Rename `.ace.example/` to `.ace-defaults/` for gem defaults directory
+
+
 ## [0.13.0] - 2025-12-30
 
 ### Changed